PointController 에 Service Layer 를 이용한 API 구현. 테스트 코드 셋업

+++ src/main/java/io/hhplus/tdd/point/PointService.java
@@ -42,7 +42,7 @@ public class PointService {
         userPointRepository.save(userId, userPoint.point() + point);
     }
 
-    public UserPoint viewPoint(int userId) {
+    public UserPoint viewPoint(Long userId) {
         return userPointRepository.findByUserId(userId);
     }
 }
